test(core): split vite hook return tests
